requirements.txt
openai-whisper>=20231117
sounddevice>=0.4.6
numpy>=1.24.0
pyyaml>=6.0
pywin32>=306
pynput>=1.7.6
pyautogui>=0.9.54
pystray>=0.19.5
Pillow>=10.0.0
torch>=2.0.0
requests>=2.31.0
tqdm>=4.66.0
win10toast>=0.6
